lcd: JZ4750: Make the frame buffer pages fully cacheable
This yields a 50% performance increase for a test program that does
nothing other than writing pixels with a simple test pattern.

To ensure pixels end up in memory to be DMA-ed, cache flushes must be
forced. This is done either at start of frame (single buffering) or on
panning (double buffering). A simple heuristic is used to figure out
whether we're running single or double buffered (see delay_flush).